# Change Dataset::CopySubrow from group-wise to column-wise