--- a/include/LightGBM/feature_group.h
+++ b/include/LightGBM/feature_group.h
@@ -232,6 +232,14 @@ class FeatureGroup {
     }
   }
 
+  inline void CopySubrowByCol(const FeatureGroup* full_feature, const data_size_t* used_indices, data_size_t num_used_indices, int fidx) {
+    if (!is_multi_val_) {
+      bin_data_->CopySubrow(full_feature->bin_data_.get(), used_indices, num_used_indices);
+    } else {
+      multi_bin_data_[fidx]->CopySubrow(full_feature->multi_bin_data_[fidx].get(), used_indices, num_used_indices);
+    }
+  }
+
   void AddFeaturesFrom(const FeatureGroup* other, int group_id) {
     CHECK(is_multi_val_);
     CHECK(other->is_multi_val_);