This is to fix the efficiency problem reported by #3637. When bagging is used, `Dataset::CopySubrow` multi-threads by feature groups, which is not quite reasonable when a large multi-value feature group exists.

This PR changes the group-wise multi-threading into column-wise multi-threading.