@guolinke 2021-01-07
any benchmark numbers?